Time: 80 ms (43.73%), Space: 16.5 MB (87.59%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.py
@@ -24,5 +24,5 @@ class Solution:
             res.append(cur_level)
         ans=[]
         for i in res:
-            ans.append(sum(i)/len(i))
+            ans.append(float(sum(i)/len(i)))
         return ans
